improve sandbox logger testability and initialization
