+++ packages/orchestrator/cmd/mock-snapshot/mock.go
@@ -117,8 +117,8 @@ func mockSnapshot(
 		IsDevelopment:    true,
 		CollectorAddress: "http://localhost:8080",
 	}
-	sbxlogger.SetSandboxLoggerInternal(ctx, loggerCfg)
-	sbxlogger.SetSandboxLoggerExternal(ctx, loggerCfg)
+	sbxlogger.SetSandboxLoggerInternal(sbxlogger.NewLogger(ctx, loggerCfg))
+	sbxlogger.SetSandboxLoggerExternal(sbxlogger.NewLogger(ctx, loggerCfg))
 
 	start := time.Now()
